move components into /datasets folder
